Fix convert type issue

+++ 2d-unet-liver-ct/data_loader.py
@@ -70,9 +70,9 @@ class DatasetHandler(Dataset):
         This function needs to be overriten from Dataset(torch) class
         """
         
-        image = imread(self.trainImages[index])
+        image = imread(self.trainImages[index]).astype(np.float32)
         label = imread(self.trainLabels[index])
-        
+	
         imageTensor, labelTensor = self.transformations(image, label)
         
         return imageTensor, labelTensor
